refactor: split omml.rs into focused sub-modules
Move the monolithic omml.rs (1,209 lines) into an omml/ directory with
focused sub-modules for better navigability and maintainability:

- omml/mod.rs: public API, OMML element parsers, and scan_math_equations
- omml/unicode.rs: unicode_to_typst() character mapping (120+ match arms)
- omml/text.rs: map_math_text() and text processing helpers

No behavior changes. All 68 omml tests pass.

Related: #138
